+++ mylog/src/constants/numbers.ts
@@ -12,6 +12,7 @@ const numbers = {
   KEYBOARD_VERTICAL_OFFSET: {
     ANDROID: -1000,
   },
+  MAX_UPLOADER_IMAGE: 5,
 } as const;
 
 const zIndex = {
@@ -19,19 +20,4 @@ const zIndex = {
   NEW_MARKER: 100,
 } as const;
 
-const months: Record<number, string> = {
-  1: 'January',
-  2: 'February',
-  3: 'March',
-  4: 'April',
-  5: 'May',
-  6: 'June',
-  7: 'July',
-  8: 'August',
-  9: 'September',
-  10: 'October',
-  11: 'November',
-  12: 'December',
-} as const;
-
-export {numbers, zIndex, months};
+export {numbers, zIndex};
